test(lib): move "test load all yaml themes" to root
as it is not just testing "YAMLTheme" or "ThemeColors"
